+++ packages/cli/src/consola/handlePromptCancel.ts
@@ -16,7 +16,7 @@
 
 import { consola } from "consola";
 
-export function handlePromptCancel(promptResponse: any) {
+export function handlePromptCancel(promptResponse: any): void {
   const isFalse = typeof promptResponse === "boolean" && !promptResponse;
   // https://github.com/unjs/consola/issues/251
   const isSigInt = typeof promptResponse === "symbol"