+++ packages/e2e.sandbox.catchall/src/runGeoQueriesTest.ts
@@ -15,12 +15,25 @@
  */
 
 import {
+  createAndFetchTempObjectSetRid,
+  hydrateObjectSetFromRid,
+} from "@osdk/client/internal";
+import {
   BoundariesUsState,
   WeatherStation,
 } from "@osdk/e2e.generated.catchall";
 import { client } from "./client.js";
 
-export async function runGeoQueriesTest() {
+const b = await createAndFetchTempObjectSetRid(client, client(WeatherStation));
+console.log(b);
+const a = await hydrateObjectSetFromRid(
+  client,
+  WeatherStation,
+  b,
+).fetchPage();
+
+console.log(a.data);
+export async function runGeoQueriesTest(): Promise<void> {
   const result = await client(WeatherStation).where({
     geohash: {
       $within: {