+++ packages/e2e.sandbox.catchall/src/runSubscriptionsTest.ts
@@ -25,7 +25,7 @@ import {
 } from "@osdk/e2e.generated.catchall";
 import { client, dsClient } from "./client.js";
 
-export async function runSubscriptionsTest() {
+export async function runSubscriptionsTest(): Promise<void> {
   let counter = 0;
   const subscription = client(OsdkTestObject)
     .subscribe(