+++ packages/e2e.sandbox.oauth.public.react-router/src/AuthenticatedRoute.tsx
@@ -6,12 +6,15 @@ import { publicOauthClient } from "./client.js";
  * A component that can be used to wrap routes that require authentication.
  * Nested routes may assume that a valid token is present.
  */
-function AuthenticatedRoute() {
+function AuthenticatedRoute(): React.JSX.Element | null {
   const [token, setToken] = useState(publicOauthClient.getTokenOrUndefined());
 
   useEffect(() => {
     publicOauthClient().then((token) => {
       setToken(token);
+    }).catch((e: unknown) => {
+      console.error(e);
+      setToken(undefined);
     });
   }, []);