+++ packages/generator/src/v2.0/generateRootIndexTsFile.ts
@@ -21,8 +21,8 @@ import { formatTs } from "../util/test/formatTs.js";
 
 export async function generateRootIndexTsFile(
   { fs, outDir, importExt, ontologyApiNamespace, ontology }: GenerateContext,
-) {
-  fs.writeFile(
+): Promise<void> {
+  await fs.writeFile(
     path.join(outDir, "index.ts"),
     await formatTs(
       `export {${