+++ packages/widget.vite-plugin.unstable/src/plugin.ts
@@ -38,7 +38,7 @@ import type {
 } from "vite";
 import { PALANTIR_PATH, SETUP_PATH, VITE_INJECTIONS } from "./constants.js";
 
-export const DIR_DIST = typeof __dirname !== "undefined"
+export const DIR_DIST: string = typeof __dirname !== "undefined"
   ? __dirname
   : path.dirname(fileURLToPath(import.meta.url));
 export interface Options {}
@@ -248,7 +248,7 @@ export function FoundryWidgetVitePlugin(_options: Options = {}): Plugin {
                 res.statusCode = settingsResponse.status;
                 res.statusMessage =
                   `Unable to set widget settings in Foundry: ${settingsResponse.statusText}`;
-                settingsResponse.text().then((err) => {
+                await settingsResponse.text().then((err) => {
                   config?.logger.error(err);
                   res.end();
                 });
@@ -669,6 +669,7 @@ function getPluginTransformHook(
     typeof plugin.transformIndexHtml === "object"
     && "transform" in plugin.transformIndexHtml
   ) {
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
     return plugin.transformIndexHtml.transform;
   }
 }