+++ packages/shared.test/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @osdk/shared.test
 
+## 2.1.0-beta.22
+
+### Minor Changes
+
+- c9da74c: Improves code quality by fixing linting errors
+
+### Patch Changes
+
+- Updated dependencies [73fa373]
+  - @osdk/api@2.1.0-beta.22
+
+## 2.1.0-beta.21
+
+### Minor Changes
+
+- f34b3d4: Add ability to fetch by rid, experimentally.
+
+### Patch Changes
+
+- Updated dependencies [3ea2e79]
+- Updated dependencies [656f975]
+- Updated dependencies [f34b3d4]
+  - @osdk/api@2.1.0-beta.21
+
 ## 2.1.0-beta.20
 
 ### Minor Changes